refactor: move bhandara-checkin-slice to slices, remove demo story as APIs are not available in Storybook
