+++ src/widgets/BhandaraCheckin/store/actions/startCheckin/startCheckinMobileOrEmailUser.ts
@@ -2,7 +2,7 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 import { ThunkApiConfig } from "../../index";
 import { getUserDetailsForEmailOrMobile } from "../utils";
 import { updateDetailsSectionSlice } from "../../slices/updateDetailsSectionSlice";
-import { bhandaraCheckinSlice } from "../../../BhandaraCheckin/bhandaraCheckinSlice";
+import { bhandaraCheckinSlice } from "../../slices/bhandaraCheckinSlice";
 
 export const startCheckinMobileOrEmailUser = createAsyncThunk<
   boolean,
